sources not compile.. extending proxy with get_user_info method. scratch
