+++ mch/session.go
@@ -1,0 +1,6 @@
+package mch
+
+type MchSession struct {
+	Config *MchConfig
+	Token  *MchToken
+}
